make distcheck has a limit on file name length, I make a separate commit for this one
